+++ code/nel/src/3d/track_sampled_vector.cpp
@@ -68,7 +68,7 @@ void	CTrackSampledVector::build(const std::vector<uint16> &timeList, const std::
 	uint i;
 
 	// reset.
-	uint	numKeys= keyList.size();
+	uint	numKeys= (uint)keyList.size();
 	_Keys.clear();
 	_TimeBlocks.clear();
 
@@ -134,7 +134,7 @@ void CTrackSampledVector::applySampleDivisor(uint sampleDivisor)
 
 	// **** rebuild the keys
 	NLMISC::CObjectVector<CVector, false>		newKeys;
-	newKeys.resize(keepKeys.size());
+	newKeys.resize((uint32)keepKeys.size());
 	for(uint i=0;i<newKeys.size();i++)
 	{
 		newKeys[i]= _Keys[keepKeys[i]];
